backend/requirements.txt
celery==5.1.0
configparser==5.0.2
Django==3.1.7
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-haystack==3.0
django-js-asset==1.2.2
django-redis==4.12.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-extensions==0.7.0
drf-haystack==1.8.10
Pillow==8.2.0
py3Fdfs==2.2.0
PyJWT==1.7.1
PyMySQL==1.0.2
redis==3.5.3
requests==2.25.1
urllib3==1.26.3
coreapi==2.3.3
mutagen==1.45.1
requests==2.25.1
openpyxl==3.0.7
django_crontab==0.7.1
DingTalkLoginTool==0.0.2
xmindparser==1.0.9
pycryptodomex==3.10.1
channels==3.0.4
asgiref==3.4.1
